Refactor non-rated issues
